refactor(blorktools): checkMeshHasHtmlContent generalization
